Move function summary types into pointer_flow
